剛体の更新と結果取得を必要なものだけに限定

+++ pkg/mwidget/model_draw.go
@@ -129,36 +129,29 @@ func updatePhysics(
 			continue
 		}
 
-		boneTransform := mbt.NewBtTransform()
-		defer mbt.DeleteBtTransform(boneTransform)
-		// if r.CorrectPhysicsType == pmx.PHYSICS_TYPE_DYNAMIC_BONE {
-		// 	mat := boneDeltas.Get(rigidBodyBone.Index).GlobalMatrix()
-		// 	bonePhysicsGlobalMatrix := r.GetRigidBodyBoneMatrix(modelPhysics)
-
-		// 	bonePhysicsGlobalMatrix[3] = mat[3]
-		// 	bonePhysicsGlobalMatrix[7] = mat[7]
-		// 	bonePhysicsGlobalMatrix[11] = mat[11]
-		// } else {
-		mat := boneDeltas.Get(rigidBodyBone.Index).GlobalMatrix().GL()
-		boneTransform.SetFromOpenGLMatrix(&mat[0])
-		// }
-
 		// 物理フラグが落ちている場合があるので、強制的に起こす
 		forceUpdate := rigidBody.UpdateFlags(model.Index, modelPhysics, enablePhysics, resetPhysics)
-		rigidBody.UpdateTransform(model.Index, modelPhysics, rigidBodyBone, boneTransform,
-			timeStep == 0.0 || !enablePhysics || forceUpdate)
+		forceUpdate = timeStep == 0.0 || !enablePhysics || forceUpdate
 
-		// mlog.Memory(fmt.Sprintf("[%d] updatePhysics[2][%d]", frame, rigidBody.Index))
+		if rigidBody.CorrectPhysicsType != pmx.PHYSICS_TYPE_DYNAMIC || forceUpdate {
+			// ボーン追従剛体・物理＋ボーン位置もしくは強制更新の場合のみ剛体位置更新
+			boneTransform := mbt.NewBtTransform()
+			defer mbt.DeleteBtTransform(boneTransform)
+			mat := boneDeltas.Get(rigidBodyBone.Index).GlobalMatrix().GL()
+			boneTransform.SetFromOpenGLMatrix(&mat[0])
+
+			rigidBody.UpdateTransform(model.Index, modelPhysics, rigidBodyBone, boneTransform)
+		}
 	}
 
 	if enablePhysics || resetPhysics {
 		modelPhysics.Update(timeStep)
 
-		// 物理（=物理後ではない）剛体位置を更新
+		// 物理剛体位置を更新
 		for _, isAfterPhysics := range []bool{false, true} {
 			for i := range len(boneDeltas.Bones.LayerSortedBones[isAfterPhysics]) {
 				bone := boneDeltas.Bones.LayerSortedBones[isAfterPhysics][i]
-				if bone.RigidBody == nil {
+				if bone.RigidBody == nil || bone.RigidBody.CorrectPhysicsType == pmx.PHYSICS_TYPE_STATIC {
 					continue
 				}
 				bonePhysicsGlobalMatrix := bone.RigidBody.GetRigidBodyBoneMatrix(model.Index, modelPhysics)
